\name{snijders}
\docType{data}
\alias{snijders}
\alias{gm00143}
\alias{gm01524}
\alias{gm01535}
\alias{gm01750}
\alias{gm02948}
\alias{gm03134}
\alias{gm03563}
\alias{gm03576}
\alias{gm04435}
\alias{gm05296}
\alias{gm07081}
\alias{gm07408}
\alias{gm10315}
\alias{gm13031}
\alias{gm13330}
\title{Public CGH data of Snijders}
\description{
15 human cell strains with known karyotype
(http://locus.umdnj.edu/nigms).
Object taken from GLAD package 1.6.0. (Hupp\'e and Barillot).
}
\usage{data(snijders)}
\references{
  \item{ }{A M Snijders, N Nowak, R Segraves, S Blackwood, N Brown, J
    Conroy, G Hamilton, A K Hindle, B Huey, K Kimura, S Law, K Myambo, J
    Palmer, B Ylstra, J P Yue, J W Gray, A N Jain, D Pinkel & D G
    Albertson , Assembly of microarrays for genome-wide measurement of
    DNA copy number,
    Nature Genetics 29, pp 263 - 264 (2001) Brief Communications}
  \item{}{Philippe Hupp\'e (2005). GLAD: Gain and Loss Analysis of DNA.
    R package version 1.6.0. http://bioinfo.curie.fr}
 }
\source{\url{http://www.nature.com/ng/journal/v29/n3/suppinfo/ng754_S1.html}}
\examples{
data(snijders)
}
\keyword{datasets}
